New: Configurable log file size limit
(cherry picked from commit 813965e6a20edef2772d68eaa7646af33028425a)

Closes #10267

(cherry picked from commit abc7efabea2e2e6a852d981ec56e7152c6ed692d)
