+++ src/NzbDrone.Common/Extensions/NumberExtensions.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Globalization;
 
 namespace NzbDrone.Common.Extensions
 {
-    public static class Int64Extensions
+    public static class NumberExtensions
     {
         private static readonly string[] SizeSuffixes = { "B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB" };
 
@@ -25,6 +25,26 @@ namespace NzbDrone.Common.Extensions
             var adjustedSize = bytes / (decimal)Math.Pow(bytesInKb, mag);
 
             return string.Format(CultureInfo.InvariantCulture, "{0:n1} {1}", adjustedSize, SizeSuffixes[mag]);
+        }
+
+        public static long Megabytes(this int megabytes)
+        {
+            return Convert.ToInt64(megabytes * 1024L * 1024L);
+        }
+
+        public static long Gigabytes(this int gigabytes)
+        {
+            return Convert.ToInt64(gigabytes * 1024L * 1024L * 1024L);
+        }
+
+        public static long Megabytes(this double megabytes)
+        {
+            return Convert.ToInt64(megabytes * 1024L * 1024L);
+        }
+
+        public static long Gigabytes(this double gigabytes)
+        {
+            return Convert.ToInt64(gigabytes * 1024L * 1024L * 1024L);
         }
     }
 }
