+++ src/NzbDrone.Core/CustomFormats/Specifications/SizeSpecification.cs
@@ -1,4 +1,5 @@
 using FluentValidation;
+using NzbDrone.Common.Extensions;
 using NzbDrone.Core.Annotations;
 using NzbDrone.Core.Validation;
