+++ src/NzbDrone.Core/MediaFiles/MovieImport/Specifications/FreeSpaceSpecification.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using NLog;
 using NzbDrone.Common.Disk;
+using NzbDrone.Common.Extensions;
 using NzbDrone.Core.Configuration;
 using NzbDrone.Core.DecisionEngine;
 using NzbDrone.Core.Download;
